Add files via upload
This repository contains the code and documentation for an IoT-based water level monitoring and controlling system. The system consists of the following components:

A water level sensor that measures the water level in a tank or reservoir and sends the data to an Arduino board.
An Arduino board that processes the sensor data and sends it to a cloud platform using a Wi-Fi module.
A cloud platform that stores and displays the sensor data on a web dashboard and allows the user to control the water level remotely using a web interface or a mobile app.
A water pump that is connected to the Arduino board and can be turned on or off by the user or by the Arduino board based on the sensor data and the desired water level.
The system aims to provide a smart and efficient solution for water level management and control in various applications, such as irrigation, water supply, hydroponics, and aquaponics. The system also aims to demonstrate the use of IoT technologies for developing innovative and practical projects.